add contenful plugin
